remove fd leak, improve error reporting
There aren't really many conditions here under which we want the user to
see a stack trace to the point in the code where the error happened, or
a list of a bunch of different related errors Combine'd into one, so I
simplify things a little here and avoid using zeebo/errs outside of the
mirror management code. When we have multiple errors, we will keep the
"first" and discard the rest. That's how APT users will be accustomed to
handling problems.

Also, we weren't closing the download object when it was done, which
would cause the APT method to run out of file descriptors eventually, no
matter high you turned them up.